knowledge/scripts/requirements.txt
google-generativeai>=0.8.0
chromadb>=0.5.0
